clean up autoupdate setState err and fix setupprogress setting
