+++ src/store/useUIStore.ts
@@ -3,7 +3,7 @@ import { backgroundType, viewType } from './types.ts';
 import { Theme } from '@app/theme/types.ts';
 import { animationDarkBg, animationLightBg, setAnimationProperties } from '@app/visuals.ts';
 
-export const DIALOG_TYPES = ['logs', 'restart'] as const;
+export const DIALOG_TYPES = ['logs', 'restart', 'autoUpdate'] as const;
 type DialogTypeTuple = typeof DIALOG_TYPES;
 export type DialogType = DialogTypeTuple[number];
 
@@ -11,6 +11,7 @@ interface State {
     theme: Theme;
     background: backgroundType;
     view: viewType;
+    latestVersion?: string;
     sidebarOpen: boolean;
     showExperimental: boolean;
     showExternalDependenciesDialog: boolean;
@@ -24,6 +25,7 @@ interface Actions {
     setShowExperimental: (showExperimental: boolean) => void;
     setShowExternalDependenciesDialog: (showExternalDependenciesDialog: boolean) => void;
     setDialogToShow: (dialogToShow: State['dialogToShow']) => void;
+    setLatestVersion: (latestVersion: string) => void;
 }
 
 type UIStoreState = State & Actions;
@@ -50,4 +52,5 @@ export const useUIStore = create<UIStoreState>()((set) => ({
     setShowExperimental: (showExperimental) => set({ showExperimental }),
     setShowExternalDependenciesDialog: (showExternalDependenciesDialog) => set({ showExternalDependenciesDialog }),
     setDialogToShow: (dialogToShow) => set({ dialogToShow }),
+    setLatestVersion: (latestVersion) => set({ latestVersion }),
 }));
